make jack graph single directed
The Connections slice kept all connections, in both directions for a jack, but
we only ever need to know about connected receivers in Transmit().  In theory
removing this may further mess up the GUI but speeds things up a bit because
the main bus in chessvm has 78 connected jacks and about half are outputs.

Also make debug breakpoints break out of "run" mode.