support calling into vm sim
Add hooks to load and call into a shared library for an outside vm simulator
specified with the -v command-line argument.

eniacsim and the vm sim must agree on a "checkpoint" cycle where they can each
compute accumulator state.  A checkpoint is signaled by a pulse to the
cy.checkpoint input.  On the first checkpoint, eniacsim dumps initial
accumulator and ft state into a snapshot the vm sim can import.  Then on
subsequent checkpoints, eniacsim runs the vm sim to update its state, and
validates that it has reached the same next state.  In case of I/O, the vm
results are skipped - eniacsim will always drive I/O itself.

Presently this is just wired up for cross-validation but in the future, the vm
simulator can take over for faster simulation, or checkpoints can be saved and
restored for better debugging.